Add Documents/PowerShell/Microsoft.PowerShell_profile.ps1 Update Documents/WindowsPowerShell/Microsoft.PowerShell_profile.ps1

+++ readonly_Documents/WindowsPowerShell/Microsoft.PowerShell_profile.ps1
@@ -1,23 +1,52 @@
+
+function quit { exit }
+
+
 # Check powershell version
 if (!$PSVersionTable.PSVersion.toString().startsWith("7")) {
-clear
-    Write-Host "Powershell version $PSVersionTable.PSVersion was detected. Running pwsh instead."
+    Write-Host "Powershell version" $PSVersionTable.PSVersion "was detected. Running pwsh instead."
     pwsh
-    exit
+    Write-Host "Exiting Powershell..."
+    Stop-Process -Id $PID -Force -PassThru
 }
+
+# Checking installed packages
+$packages = @("ajeetdsouza.zoxide", "Starship.Starship", "Neovim.Neovim", "Fastfetch-cli.Fastfetch")
+$packages | ForEach-Object {
+winget list -q $_ | Out-Null
+if ($?) {
+    # Write-Host "$_ installed"
+} else {
+        Write-Host "Package $_ not found. Installing..."
+        winget install $_
+    }
+}
+
 
 # Zoxide
 Invoke-Expression (& { (zoxide init powershell | Out-String) })
-Remove-Item alias:cd
+if (test-path alias:z) {
+    Remove-Item alias:z
+}
 new-alias cd z
-echo "Zoxide loaded."
+Write-Host "Zoxide loaded."
 
 # Other aliases
-Remove-Item alias:cls
+
+if (test-path alias:cls) {
+    Remove-Item alias:cls
+}
 new-alias cls clear
-
-Remove-Item alias:q
-new-alias q exit
-
+if (test-path alias:q) {
+    Remove-Item alias:q
+}
+new-alias q quit
+if (test-path alias:qa) {
+    Remove-Item alias:qa
+}
 Remove-Item alias:qa
-new-alias qa exit
+new-alias qa quit
+
+Write-Host "========================================================"
+
+fastfetch
